put CompositeSpecification extending classes in one file
