+++ src/modules/organisation/specification/zugehoerig-zu-traeger.ts
@@ -1,7 +1,7 @@
 import { OrganisationDo } from '../domain/organisation.do.js';
 import { OrganisationRepo } from '../persistence/organisation.repo.js';
 import { OrganisationsTyp } from '../domain/organisation.enums.js';
-import { CompositeSpecification } from '../../specification/specifications.js';
+import { CompositeSpecification } from '../../specification/composite-specification.js';
 
 export class ZugehoerigZuTraeger extends CompositeSpecification<OrganisationDo<true>> {
     public constructor(private readonly organisationRepo: OrganisationRepo) {
